requirements.txt
python-telegram-bot==20.0
pydub==0.25.1
numpy==1.26.4
librosa==0.10.2.post1
soundfile==0.12.1
noisereduce==3.0.2
scipy==1.13.0
openai==1.17.0
python-dotenv==1.0.1
python-docx==0.8.11
ffmpeg-python==0.2.0  # Opzionale
